refactor: 눈꽃 에니메이션 리렌더링 이슈 해결
