+++ src/components/InputForm.js
@@ -1,17 +1,10 @@
+import { useMemo } from 'react'
 import styled from 'styled-components'
 import { Typography } from '@/components'
 import { flexSpaceBetween, flexCenter } from '@/styles/common'
 
 export default function InputForm(props) {
-  const {
-    inputValue,
-    onChange,
-    label,
-    placeholder,
-    errorMessage,
-    name,
-    onBlurInputs,
-  } = props
+  const { inputValue, onChange, label, placeholder, errorMessage, name } = props
 
   return (
     <InputLayout>
@@ -32,7 +25,6 @@ export default function InputForm(props) {
             name={name}
             onChange={(e) => onChange(e)}
             type={name === 'password' ? 'password' : 'text'}
-            onBlur={() => onBlurInputs(name, inputValue)}
           />
         </InputBox>
       </InputWrapper>
