+++ src/stores/userInfo.js
@@ -1,25 +1,5 @@
-import { ERROR_MESSAGE } from '@/lib/constants'
 import { create } from 'zustand'
 import { persist, createJSONStorage } from 'zustand/middleware'
-
-export const useUserInputStore = create((set) => ({
-  userInputValue: {
-    nickname: '',
-    password: '',
-  },
-
-  errorMessage: {
-    id: ERROR_MESSAGE.ID?.default,
-    pw: ERROR_MESSAGE.PW?.default,
-  },
-
-  setErrorMessage: (error) => set(() => ({ errorMessage: error })),
-
-  setUserInputValue: (userInfo) =>
-    set(() => ({
-      userInputValue: userInfo,
-    })),
-}))
 
 export const useBundlesStore = create((set) => ({
   currentIndex: 0,
